loveegoai-backend/requirements.txt
# FastAPI core
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
motor==3.6.0
pymongo==4.9.0
dnspython==2.7.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# AI and LLM
openai==1.59.5
pinecone-client==5.0.1

# TTS
edge-tts==6.1.18

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
python-docx==1.1.2

# Scheduler
apscheduler==3.11.0

# Logging
loguru==0.7.3
